backend/requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.5.2
bcrypt==4.3.0
blinker==1.8.2
click==8.1.8
DateTime==5.5
dnspython==2.6.1
email-validator==2.3.0
exceptiongroup==1.3.0
factory_boy==3.3.3
Faker==35.2.2
fastapi==0.117.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==2.1.5
marshmallow==3.22.0
packaging==25.0
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
pytest==8.3.5
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
SQLAlchemy-serializer==1.4.12
starlette==0.44.0
tomli==2.2.1
typing_extensions==4.13.2
Werkzeug==3.0.6
zipp==3.20.2
zope.interface==7.2
zxcvbn==4.5.0
